requirements.txt
groq
langchain
langchain-core
langchain-groq
langchain-community
pandas
python-dotenv
streamlit